Reorganized components, started on availability page
