+++ park-sense/src/components/general/MainNavBar.js
@@ -3,7 +3,7 @@ import { PrimeReactProvider } from 'primereact/api'
 import { Menubar } from 'primereact/menubar';
 import { InputText } from 'primereact/inputtext';
 import { useNavigate } from 'react-router-dom'
-import "./../../../node_modules/primereact/resources/themes/lara-light-blue/theme.css";
+import "primereact/resources/themes/lara-light-blue/theme.css";
 import 'primereact/resources/primereact.min.css';
 import logo2 from "../../img/parksense-logo2.png";
