--- park-sense/src/components/pages/Availability.js
@@ -1,5 +1,0 @@
-import WhatIsParkSense from "../assets/WhatIsParkSense";
-
-export default function Availability(){
-    return(<p>Availability Component</p>)
-}
