+++ park-sense/src/components/pages/availability/Availability.js
@@ -1,0 +1,11 @@
+import LotDropdown from "./LotDropdown"
+import LotStatus from "./LotStatus"
+
+export default function Availability(){
+    return(
+        <center>
+            <LotDropdown></LotDropdown>
+            <LotStatus></LotStatus>
+        </center>
+    );
+}
