+++ park-sense/src/components/pages/availability/LotStatus.js
@@ -1,0 +1,12 @@
+export default function LotStatus(){
+    return(
+        <div>
+            <h1>Availability:</h1>
+                <p>Spots Available: 11/100</p>
+            <br></br>
+            <h1>General Info:</h1>
+            <p>Lot Parking Pass: Yellow</p>
+            <p>Yellow Pass Requred (8am-5pm Monday-Friday)</p>
+        </div>
+    );
+}
