+++ park-sense/src/components/pages/no-page/NoPage.js
@@ -1,5 +1,3 @@
-
-
 export default function NoPage(){
     return(
         <h2>Error</h2>
