+++ park-sense/src/components/pages/parking-map/ParkingMap.js
@@ -1,5 +1,3 @@
-import WhatIsParkSense from "../assets/WhatIsParkSense";
-
 export default function ParkingMap(){
     return(<p>ParkingMap Component</p>);
 }
